@janeyx99 2022-08-22
@pytorchbot revert -m "Sorry, reverting as this breaks mac functorch tests on trunk https://hud.pytorch.org/pytorch/pytorch/commit/f56720ea7c7ad0bcb4c5af669e28bf7de8122cb6" -c nosignal

Please add a ciflow/trunk label to your next PR attempt to run additional trunk tests on your PR and verify that the mac functorch tests pass. To prevent reverts like this in the future, try out `merge -l` which will run land checks (Lint + pull + trunk checks) before merging. See https://github.com/pytorch/pytorch/wiki/Bot-commands for more details.