@CaoE 2022-08-24
> Hi @CaoE , thank you for your PR but it seems to introduce some numerical out of bounds in certain unit test cases. Some models also show accuracy issues.

Hi @weiwangmeta, could you please share the detailed info about the numerical out of bounds and accuracy ? Thank you.