@weiwangmeta 2022-08-24
@CaoE  Currently I do not have a standalone reproducer. The failed testing case was named: 

test_block_interleaved_linear_operator.TestBlockInterleavedLinearOperatorBatch
  test_solve_matrix_broadcast: AssertionError: tensor1 (torch.Size([3, 2, 24, 5])) and tensor2 (torch.Size([3, 2, 24, 5])) are not close enough. 

Failures:

  1) test_block_interleaved_linear_operator.TestBlockInterleavedLinearOperatorBatch: test_solve_matrix_broadcast
    1) AssertionError: tensor1 (torch.Size([3, 2, 24, 5])) and tensor2 (torch.Size([3, 2, 24, 5])) are not close enough. 
    max rtol: 337876528340982113272697256325974851584.00000000		max atol: 609890353376311672964835584975568896.00000000