@CaoE 2022-08-24
Hi @weiwangmeta, is test_solve_matrix_broadcast a internal UT? How can I reproduce the error? Could you help to reproduce it, thank you very much.