@weiwangmeta 2022-08-24
Yes, it is internal UT. I will try to come up with a standalone reproducer, but it will be a bit of wait.