@malfet 2022-09-20
Here is a simple reproducer. If transpose works the way it supposed to, below function should return `24, 24, 414720`, but with above-mentioned PR applied it returns garbage (for example `tensor(0.) tensor(1.6096e+22) tensor(1.1589e+24)`)
```
import torch
foo = torch.ones([2, 5, 24, 24])
i_mat = torch.ones([3, 2, 5, 24, 24])
y = foo.permute(0, 1, 3, 2).matmul(i_mat)
print(y.min(), y.max(), y.sum())
```

Some notes on how to get a reproducer: `torch.save` is a friend :) And then just try to minimize it...