@CaoE 2022-09-22
> Please consider the following for the PR description: 1) the motivation of PR. which model (like word2vec) could be benefited, target platform, and use case. 2) existing related optimization PRs. 3) the limitation (scope) of the optimization, like you only target bf16, last 2 dimension, and need to have certain size. 4) the technique you use. like parallel processing of decomposed small blocks of data, and each block processed with the brgemm's SIMD transpose fuction.

Thank you @Jianhui-Li for your detailed advice ! I will follow the suggestions and improve the PR.